requirements.txt
pypdf>=4.0.0
cryptography>=3.1
tkinterdnd2>=0.3.0
python-docx>=1.1.0
markdown>=3.5.0
pdf2docx>=0.5.0
pymupdf4llm>=0.0.5
